Refactor ToDo class to include a property for the type
This commit refactors the ToDo class in the notion_api/notion_client_wrapper/block/to_do.py file. It adds a new property called "type" that returns the string "to_do". This enhancement improves the clarity and consistency of the code by explicitly defining the type of the ToDo block.