feat: Add TaskContextType property to ToDoTask
This commit adds a new property `context` to the `ToDoTask` class in the `task.py` file. The `context` property returns a list of `TaskContextType` objects, representing the context of the task. This allows for better organization and categorization of tasks based on their context. The `context` property is populated by retrieving the multi-select property named "コンテクスト" and converting the values to `TaskContextType` objects. This enhancement improves the functionality and flexibility of the `ToDoTask` class.