feat: Add upload_as_file method to SlackClient
This commit adds the `upload_as_file` method to the `SlackClient` class in the `slack_client.py` file. The `upload_as_file` method allows for uploading a text content as a file to Slack. This enhancement improves the functionality of the `SlackClient` class by providing a convenient way to share text content as files in Slack channels.