feat: Add support for Webclip page type in InboxService
This commit updates the InboxService class in the inbox_service.py file to add support for the Webclip page type. It includes a new condition in the get_page_type method to check if the page is an instance of the Webclip class and returns the corresponding page type string "【Webクリップ】". This enhancement improves the functionality of the InboxService by correctly identifying and categorizing Webclip pages.