chore: Update VSCode extensions and settings
This commit updates the VSCode extensions and settings. It adds the "getgauge.gauge" extension to the extensions.json file and includes new settings in the settings.json file. The new settings include associations for "*.spec" and "*.cpt" files to the "gauge" language, as well as enabling auto-save with a delay of 500 milliseconds. These updates improve the functionality and customization options of the VSCode environment.