feat: Add support for gauge-current tag in e2e tests
This commit adds support for the `gauge-current` tag in the e2e tests. It introduces a new `gauge-current` target in the Makefile, allowing developers to run only the tests tagged with `current`. This enhancement improves the test execution process by providing a way to selectively run tests related to the current functionality being developed or tested.